% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic-sdg.R
\name{writeToFile}
\alias{writeToFile}
\title{Write miicsdg files to a new folder. The function will create the folder you specified. These two files can be loaded at the page: https://miic.curie.fr/vis_NL.php}
\usage{
writeToFile(miicsdg_res, folder)
}
\arguments{
\item{miicsdg_res}{the output of the miicddg algorithm}

\item{folder}{A new folder where to store the data}
}
\description{
Write miicsdg files to a new folder. The function will create the folder you specified. These two files can be loaded at the page: https://miic.curie.fr/vis_NL.php
}
